# docs/openapi.yaml
openapi: 3.0.3
info:
  title: VideoForge AI API
  description: |
    Complete API for VideoForge AI - Automated Video Generation Platform
    
    ## Authentication
    All endpoints (except /api/health) require authentication via Supabase JWT token.
    Include the token in the Authorization header: `Bearer <token>`
    
    ## Rate Limits
    - Free tier: 10 requests/minute
    - Starter: 30 requests/minute
    - Pro: 100 requests/minute
    - Business: 300 requests/minute
    
    ## Video Generation Pipeline
    1. Create a project with topic and settings
    2. Call POST /api/projects/{id}/generate to start pipeline
    3. Poll GET /api/projects/{id}/generate for status
    4. Once complete, publish via POST /api/projects/{id}/publish
  version: 1.0.0
  contact:
    name: VideoForge AI Support
    email: support@videoforge.ai
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.videoforge.ai
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login
      description: Authenticate with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  example: "SecurePass123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        type: object
                        properties:
                          id:
                            type: string
                          email:
                            type: string
                          emailVerified:
                            type: boolean
                          createdAt:
                            type: string
                            format: date-time
                      session:
                        type: object
                        properties:
                          accessToken:
                            type: string
                            description: JWT access token for API calls
                          refreshToken:
                            type: string
                            description: Token to refresh access token
                          expiresAt:
                            type: integer
                            description: Unix timestamp when token expires
                          expiresIn:
                            type: integer
                            description: Seconds until token expires
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh token
      description: Get a new access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  description: Refresh token from login/signup response
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      session:
                        type: object
                        properties:
                          accessToken:
                            type: string
                          refreshToken:
                            type: string
                          expiresAt:
                            type: integer
                          expiresIn:
                            type: integer
                      user:
                        type: object
                        properties:
                          id:
                            type: string
                          email:
                            type: string
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Forgot password
      description: Send password reset email. Always returns success to prevent email enumeration.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
      responses:
        '200':
          description: Reset email sent (if account exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "If an account exists with this email, you will receive a password reset link"
...